minor reorganization
