+++ megamek/src/megamek/server/victory/Victory.java
@@ -13,6 +13,7 @@
  */
 package megamek.server.victory;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -22,8 +23,9 @@ import megamek.common.Report;
 import megamek.common.options.GameOptions;
 import megamek.common.options.OptionsConstants;
 
-public class Victory {
-
+public class Victory implements Serializable {
+    private static final long serialVersionUID = -8633873540471130320L;
+    
     private boolean checkForVictory;
     private int neededVictoryConditions;
